Playback Rate
ardcode is the code meant to be uploaded on to an arduino(in this case i'm using a nano) connected to an external potentiometer (for testing purposes).
testingprototype contains a python function that gets the serial output from the arduino.
thirdprototype contains the main code. its currently in testing so it wont function as a normal music player.