Getting GPS Data From android
this is a different approach as compared to the previous one. Enable debugging mode on android and have it connected to the system via usb cable. please do turn on google maps on your phone as the speed data sent to the code will be the last known speed recorded.

the ffts in this code also differ from the previous ones as this one uses a vocal seperation which doesnt work well, and may distort the audio if you go beyond the set speed. you can change the max speed by changing the denominator value in proton.py.